renaming of blood_type_field
